Functional Commit
This completes the initial changes to sever spdm-lib from its dependency on
libcaliptra.

Added an implementation of SpdmHash to the emulator application. This
implementation isn't clean, and causes a loss of error information, but
is a working proof-of-concept. Currently passes the CI test which uses
libspdm and spdm-emu to assert validity.

Currently this change only replaces the hashing step for GET_DIGEST,
but should also work for managing the transcript, which will be the next
change.

This pattern should be trivial for SpdmTransport, as well as cryptographic
services.