rom: Add stub DOT implementation and MCI FW fatal error (#623)
* rom: Add stub DOT implementation and MCI FW fatal error

This adds a simple DOT implementation in the ROM that always fails if
the DOT is not empty.

Mostly this is implemented so that we can get all of the other tests and
plumbing needed:

* Add `FlashStorage` impl in the ROM to support reading and writing the
  DOT blob
* Add a small memory in the emulator for the DOT flash
* Adds ability for tests to write to the DOT flash
* Sets the MCI FW fatal error register when a fatal error occurs
* Writes a simple test that a corrupt DOT flash causes an error

There appears to be a possible compiler bug in the `copy_from_slice()`
operation, which is noted in the `SimpleFlash` struct: even though the
lengths are guaranteed to be equal, the compiler will add a panic check.
Even if I add an `assert_eq!` with the two length, the compiler seems to
add the panic. Temporarily working around this using
`ptr::copy_nonverlapping()`, which is what `copy_from_slice` simplifies
to.

* Add fatal error set in fpga platform

Cleans up missing impls and clippy warnings.

* Add docs for main fuses

* Use random port each time for I3C test; reset MCU_RUNNING at beginning of each test