Switch to nexttest for running emulator test (#662)
For some reason, the test_soc_boot is not being run by CI. Currently CI uses 'cargo test' to run
the unit and integration tests.
Switching to 'cargo nexttest run' was able to solve this problem and run all tests.

It was also found that ROM isn't compiling properly due to the ROM code size is above 32KB. To fix this,
the ROM_SIZE was increased to 64 KB.

* Use a different profile for emulator tests since they are slower