Set FW_EXEC_CTRL bit of SoC after loading image (#671)
Fixes: https://github.com/chipsalliance/caliptra-mcu-sw/issues/508

Set the FW_EXEC_CTRL of the the SoC after loading its image. This is
done by sending the ACTIVATE_FIRMWARE mailbox command to Caliptra.
Since the SoC architecture is platform dependent, there could be
additional integrator steps to activate the SoC. Therefore, this is just
an example of how the SoC can be activated.